added the rum
Added the function to remove the rum and add points once player goes over the rum.
I have noticed that I may not need to repeat the add and remove function for each element. Will try to create 1 function that can pass through the element and position in the perameter